+++ packages/core/src/internals.ts
@@ -61,6 +61,10 @@ import {
 import * as CSSUtils from './default-css';
 import {FEATURE_FLAG_FIREFOX_SUPPORT} from './feature-flags';
 import {getRemotionEnvironment, RemotionEnvironment} from './get-environment';
+import {
+	INITIAL_FRAME_LOCAL_STORAGE_KEY,
+	setupInitialFrame,
+} from './initial-frame';
 import {isAudioCodec} from './is-audio-codec';
 import * as perf from './perf';
 import {
@@ -74,9 +78,7 @@ import {SequenceContext} from './sequencing';
 import {
 	ENV_VARIABLES_ENV_NAME,
 	ENV_VARIABLES_LOCAL_STORAGE_KEY,
-	INITIAL_FRAME_LOCAL_STORAGE_KEY,
 	setupEnvVariables,
-	setupInitialFrame,
 } from './setup-env-variables';
 import * as Timeline from './timeline-position-state';
 import {
